Add test-includes.sh and includes.txt
includes.txt is intended to list functions which we expect to return
True when run through graph-refine. More explicitly, it is the list of
functions which test-includes.sh will pass to graph-refine for testing.

test-includes.sh will run graph-refine on the functions listed in
includes.txt, on the data stored in seL4-example/latest.

Signed-off-by: Mitchell Buckley <mitchell.buckley@data61.csiro.au>