Add make_solverlist.sh.
This script can be used to automatically generate a solver list that
will be accepted by graph-refine. Its current behaviour reflects my
understanding of the settings we require, but this commit should not be
taken as an authority.

By writing a separate script for this task I hope to modularise any
build system for graph-refine and simplify regression testing.

Signed-off-by: Mitchell Buckley <mitchell.buckley@data61.csiro.au>