Refactor logging and argument parsing
Previous versions of `graph-refine` parsed the `trace-to:`
argument after loading `ASMFunctions.txt` and `CFunctions.txt`.
This meant that we could not write to the log/report until these
files were loaded.

I refactor argument parsing and logging to make the order of events
a little more sensible. This allows us to enable logging while
loading the target files.

Signed-off-by: Zoltan Kocsis <zoltan.kocsis@data61.csiro.au>