word64: restore uniform token_smt_typ across architectures
The `token_smt_typ` was set to the architecture word size in
`solver.py`. This was due to a misleading comment in an
earlier version of the code base. The type should be 64-bit
word, uniformly across all architectures in order for it
to corresponds to the actual behavior of the old, arm7-only
branch.

This patch sets `token_smt_typ` to word64T.

Signed-off-by: Zoltan A. Kocsis <zoltan.kocsis@data61.csiro.au>